remove sources for AviUtl1
