+++ CHANGELOG.md
@@ -76,32 +76,6 @@
 - 倍角化を削除．
 - 内部の軽微な修正．
 
-## 1.3.1 2026/02/05
-- カスタムオブジェクトにおいて，拡大率を変えるとアンカーポイントと描画がズレるバグを修正．
-- 折れ線矢印の描画を調整．
-
-## 1.3.2 2026/03/02
-- 内部の軽微な修正．
-
-## 1.4.0 2026/03/18
-- スレッドプール方式にし，スレッド数を指定するパラメータを削除．
-
-## 1.4.1 2026/03/23
-- 完全透明のピクセルを含むオブジェクトに画素平均法リサイズを適用した場合に，ゼロ除算でクラッシュするバグを修正．
-
-## 1.4.2 2026/04/02
-- 丸めを四捨五入から `std::nearbyint` 及び明示的な最近接偶数丸めに変更．
-- 累積的なコーディングの修正．
-
-## 1.4.3 2026/04/03
-- エフェクトを使わなかった場合に，本体終了時にクラッシュするバグを修正．
-
-## 1.4.4 2026/04/03
-- クラッシュすることがあるバグを修正．
-
-## 1.4.5 2026/04/08
-- 軽微なコーディングの修正．
-
 ## 2.0.0 2026/02/05
 - AviUtl ExEdit2 用を追加．可能なパラメータの設定法が異なるため，無印用とは一部インターフェースが異なる．
 
@@ -132,3 +106,7 @@
 
 ## 2.1.6 2026/04/08
 - 軽微なコーディングの修正．
+
+## 2.1.7 2026/04/22
+- AviUtl 無印用をブランチ分離して AviUtl ExEdit2 専用に．
+- 累積的なコーディングの修正．
