+++ USAGE.md
@@ -1,7 +1,6 @@
-# @ksa.obj
-カスタムオブジェクト用のスクリプトです．
+# カスタムオブジェクト用スクリプト
 
-## 折れ線矢印
+## KSA折れ線矢印.obj2
 本家の「ライン(移動軌跡)」のように長さを時間変化させるための折れ線矢印です．長さは折れ線全体の長さに対するパーセンテージを「描画率」のパラメータで調整します．「ライン(移動軌跡)」が鏃の中心を軸にしているのに対し，「折れ線矢印」は鏃の先端を軸にしているため，特定の座標を指したいときに使いやすいです．代わりに，折れる途中の描画が手抜きで汚いです．
 
 <dl>
@@ -21,7 +20,7 @@
   <dd><code>{頂点1のx, 頂点1のy, 頂点2のx, ...}</code>の形で座標を指定します．アンカーポイントで動かす場合は気にしなくて大丈夫です．</dd>
 </dl>
 
-## 両矢印
+## KSA両矢印.obj2
 両側に鏃のある単純な矢印です．
 
 <dl>
@@ -37,14 +36,14 @@
   <dd><code>{頂点1のx, 頂点1のy, 頂点2のx, 頂点2のy}</code>の形で座標を指定します．アンカーポイントで動かす場合は気にしなくて大丈夫です．</dd>
 </dl>
 
-## 矩形枠
+## KSA矩形枠.obj2
 「角度」が`0`の場合に，各辺がx軸またはy軸に平行な長方形の枠です．
 
 <dl>
  <dt>線幅</dt>
   <dd>線の幅をピクセル数で指定します．</dd>
  <dt>角丸</dt>
-  <dd><code>1</code>なら角を丸くし，<code>0</code>なら尖らせます．丸みの半径は線幅の半分で固定です．</dd>
+  <dd>チェックなら角を丸くし，未チェックなら尖らせます．丸みの半径は線幅の半分で固定です．</dd>
  <dt>アンカー数</dt>
   <dd><code>2</code>なら対角線上の2頂点を，<code>1</code>なら1頂点をアンカーポイントとします．角度を変える場合は<code>1</code>にするとアンカーポイントと描画頂点がずれません．</dd>
  <dt>色</dt>
@@ -53,7 +52,7 @@
   <dd>アンカーポイントの座標が出力されます．数字を指定して頂点位置を決めてもよいです．</dd>
 </dl>
 
-## 楕円枠
+## KSA楕円枠.obj2
 アンカーポイントで生成される長方形に沿った楕円形の枠です．
 
 <dl>
@@ -71,8 +70,7 @@
   <dd>アンカーポイントの座標が出力されます．数字を指定して頂点位置を決めてもよいです．</dd>
 </dl>
 
-
-## 枠付き角丸四角形
+## KSA枠付き角丸四角形.obj2
 角を丸められ，枠と塗りつぶし領域で独立に色や透明度を調整できる長方形オブジェクトです．
 
 <dl>
@@ -94,10 +92,9 @@
   <dd>アンカーポイントの座標が出力されます．アンカーポイントは外径を決定します．</dd>
 </dl>
 
-# @ksa.anm
-アニメーション効果用のスクリプトです．
+# アニメーション効果用スクリプト
 
-## 透明度グラデーション
+## KSA透明度グラデーション.anm2
 高透明度側と低透明度側の透明度を任意に指定できる透明度グラデーションです．
 
 <dl>
@@ -113,21 +110,21 @@
   <dd><code>0</code>なら線形状，<code>1</code>ならsinカーブ状のグラデーションを掛けます．</dd>
 </dl>
 
-## 縁透明グラデーション / KSA縁透明G
+## KSA縁透明G.anm2
 長方形のオブジェクトの縁に透明度のグラデーションを掛け，背景に馴染むようにする効果です．
 
 <dl>
  <dt>幅</dt>
   <dd>縁から何 px に亘ってグラデーションを掛けるかを指定します．</dd>
  <dt>角丸</dt>
-  <dd><code>1</code>またはチェックしているなら角を丸くします．</dd>
+  <dd>チェックしているなら角を丸くします．</dd>
  <dt>タイプ</dt>
   <dd><code>0</code>なら線形状，<code>1</code>なら円状のグラデーションを掛けます．</dd>
  <dt>追加上，追加下，追加左，追加右</dt>
   <dd>上下左右の幅を追加します．向きによって異なる幅にしたいときに指定します．</dd>
 </dl>
 
-## 回転
+## KSA回転.anm2
 「広がって登場」をずっと繰り返すようなアニメーションです．
 仮想バッファ利用がうまく行かず，直接描画しているので，最後のエフェクトにする必要があります．
 
@@ -140,7 +137,7 @@
   <dd>回転を逆方向にします．</dd>
 </dl>
 
-## Lanczos3リサイズ / KSAリサイズ
+## KSAリサイズ.anm2
 Lanczos3 関数を使ったリサイジンクを行います．
 マルチスレッド処理をしますが，それでも非常に重いので，使用時はご留意ください．
 
@@ -155,7 +152,7 @@ Lanczos3 関数を使ったリサイジンクを行います．
   <dd>リサイズ前に指定したピクセル数のクリッピングを行います．</dd>
 </dl>
 
-## インターレース解除
+## KSAインターレース解除.anm2
 インターレース解除をしながら動画を読み込むオブジェクトです．
 元のオブジェクトを無視して，指定したパスの動画のインターレース解除画像を出力します．
 正常に動作させるには，プロジェクトのフレームレートを入力動画の 2 倍にする必要があります．
@@ -180,7 +177,7 @@ Lanczos3 関数を使ったリサイジンクを行います．
 
 240p の疑似インターレース動画には`1`を使用します．480i などのインターレース動画では，チラツキはあるがゴーストのない`2`か，ゴーストが出てぼやけるが滑らかな`4`を使用します．
 
-## 加減速入退場
+## KSA加減速入退場.anm2
 減速しながら入場し，加速しながら退場するアニメーションです．
 
 <dl>
@@ -190,7 +187,7 @@ Lanczos3 関数を使ったリサイジンクを行います．
   <dd>入場前，退場後の相対位置を指定します．</dd>
 </dl>
 
-## 点滅入退場
+## KSA点滅入退場.anm2
 点滅しながら入退場するアニメーションです．
 
 <dl>
@@ -200,24 +197,23 @@ Lanczos3 関数を使ったリサイジンクを行います．
   <dd>点(表示)長さと滅(非表示)長さをフレーム数で指定します．</dd>
 </dl>
 
-## スクロール
+## KSAスクロール.anm2
 表示領域を固定したまま，縦または横にスクロールさせるようなアニメーションです．
 
 <dl>
  <dt>再生位置</dt>
   <dd><code>0</code>が上端または左端，<code>100</code>が下端または右端に対応し，これをトラックバー変化させることでスクロールします．</dd>
  <dt>縦/横</dt>
-  <dd><code>0</code>または未チェックなら縦スクロール，<code>1</code>またはチェックなら横スクロールします．</dd>
+  <dd>未チェックなら縦スクロール，チェックなら横スクロールします．</dd>
  <dt>表示長</dt>
   <dd>表示させる領域のサイズ(縦なら高さ，横なら幅)を指定します．オブジェクトのサイズより大きい場合は何もしません．</dd>
  <dt>拡大</dt>
-  <dd><code>1</code>またはチェックなら領域拡張を行い，再生位置<code>0</code>および<code>100</code>がなにも見えない状態になります．</dd>
+  <dd>チェックなら領域拡張を行い，再生位置<code>0</code>および<code>100</code>がなにも見えない状態になります．</dd>
 </dl>
 
-# @ksa.tra
-トラックバー変化方法用のスクリプトです．
+# トラックバー変化方法用スクリプト
 
-## 加減速移動
+## KSA加減速移動.tra2
 本家の加減速移動がバグっているので追加した加減速移動スクリプト．
 パラメータの値 (デフォルト：`5`) によって，次の式に比例した形で変化させます．
