--- ksa/@ksa.anm
@@ -1,236 +1,0 @@
-@透明度グラデーション
---track0:中心X,-10000,10000,0,1
---track1:中心Y,-10000,10000,0,1
---track2:角度,-360,360,0
---track3:幅,1,1000,100,1
---dialog:開始透明度,local s_tp=0;終了透明度,local e_tp=100;タイプ,local type=0;
-
-local ksa = require("ksa")
-
-local s_alp = (1-s_tp/100)
-local e_alp = (1-e_tp/100)
-local angle = math.rad( obj.track2 )
-
-local data, w, h = obj.getpixeldata()
-ksa.trsgrad(data, w, h, obj.track0+obj.cx+w/2, obj.track1+obj.cy+h/2, angle, obj.track3, s_alp, e_alp, type)
-obj.putpixeldata(data)
-
-@縁透明グラデーション
---track0:幅,0,1000,10,1
---track1:角丸,0,1,1,1
---track2:タイプ,0,1,1,1
---dialog:追加上,local pt=0;追加下,local pb=0;追加左,local pl=0;追加右,local pr=0;
-
-local ksa = require("ksa")
-
-local data, w, h = obj.getpixeldata()
-local gw = obj.track0;
-ksa.edgegrad(data, w, h, gw+pt, gw+pb, gw+pl, gw+pr, obj.track1, obj.track2)
-obj.putpixeldata(data)
-
-@回転
---track0:回転量,0,100,6,1
---track1:縦回転,0,1,0,1
---track2:反転,0,1,0,1
-
-local ksa = require("ksa")
-
-local angle = math.pi*obj.track0*obj.frame/obj.totalframe
-if ( obj.track2 == 1 ) then
-	angle = -angle
-end
-local x0, x1, y0, y1
-if ( obj.track1 == 0 ) then
-	local dw = obj.w*math.sin(angle)
-	x0 = -dw/2
-	x1 = dw/2
-	y0 = -obj.h/2
-	y1 = obj.h/2
-else
-	local dh = obj.h*math.sin(angle)
-	x0 = -obj.w/2
-	x1 = obj.w/2
-	y0 = -dh/2
-	y1 = dh/2
-end
-
-ksa.drawrect(x0, y0, x1, y1)
-
-@Lanczos3リサイズ
---track0:拡大率,0,800,100,0.01
---track1:X,0,10000,100,0.01
---track2:Y,0,10000,100,0.01
---track3:平均法,0,1,0,1
---check0:ドット数でサイズ指定,0
---dialog:上,local ct=0;下,local cb=0;左,local cl=0;右,local cr=0;
-
-local ksa = require("ksa")
-local src, sw, sh = obj.getpixeldata("alloc")
-local dw, dh
-if ( obj.check0 ) then
-	dw = ksa.round(obj.track1)
-	dh = ksa.round(obj.track2)
-else
-	dw = ksa.round( ( sw-cl-cr )*obj.track0*obj.track1*1e-4 )
-	dh = ksa.round( ( sh-ct-cb )*obj.track0*obj.track2*1e-4 )
-end
-if ( ( cl+cr < sw ) and ( ct+cb < sh ) and ( 0 < dw ) and ( 0 < dh ) ) then
-	obj.setoption("drawtarget", "tempbuffer", dw, dh)
-	obj.copybuffer("obj", "tmp")
-	local dest, _dw, _dh = obj.getpixeldata()
-	if ( _dw < dw ) then
-		local clr = math.ceil( (dw-_dw)*(sw-cl-cr)/(dw*2) )
-		cl = cl+clr
-		cr = cr+clr
-		if ( obj.check0 ) then
-			dw = ksa.round( dw*(sw-cl-cr)/(sw-cl-cr+clr*2) )
-		else
-			dw = ksa.round( ( sw-cl-cr )*obj.track0*obj.track1*1e-4 )
-		end
-	end
-	if ( _dh < dh ) then
-		local ctb = math.ceil( (dh-_dh)*(sh-ct-cb)/(dh*2) )
-		ct = ct+ctb
-		cb = cb+ctb
-		if ( obj.check0 ) then
-			dh = ksa.round( dh*(sh-ct-cb)/(sh-ct-cb+ctb*2) )
-		else
-			dh = ksa.round( ( sh-ct-cb )*obj.track0*obj.track2*1e-4 )
-		end
-	end
-	if ( dw ~= _dw or dh ~= _dh ) then
-		obj.setoption("drawtarget", "tempbuffer", dw, dh)
-		obj.copybuffer("obj", "tmp")
-		dest, dw, dh = obj.getpixeldata()
-	end
-	if ( obj.track3 == 0 ) then
-		ksa.clip_resize(src, sw, sh, dest, dw, dh, ct, cb, cl, cr)
-	else
-		ksa.clip_resize_ave(src, sw, sh, dest, dw, dh, ct, cb, cl, cr)
-	end
-	obj.putpixeldata(dest)
-else
-	obj.setoption("drawtarget", "tempbuffer", 1, 1)
-	obj.copybuffer("obj", "tmp")
-end
-
-@インターレース解除
---track0:解除法,0,4,2,1
---track1:BFF/TFF,0,1,1,1
---file:
-
-local ksa = require("ksa")
-obj.load("movie", file, (obj.frame-0.5)/obj.framerate)
-local dest, w, h = obj.getpixeldata("alloc")
-
-if ( obj.track0 == 1 ) then -- 無補間
-	ksa.deinterlace_nn(dest, w, h, XOR(obj.frame%2, obj.track1))
-elseif ( obj.track0 == 2 ) then -- 空間方向補間
-	ksa.deinterlace_spatial(dest, w, h, XOR(obj.frame%2, obj.track1))
-elseif ( obj.track0 == 3 ) then -- 時間方向補間
-	obj.load("movie", file, (obj.frame-0.5-1)/obj.framerate)
-	local bf, _w, _h = obj.getpixeldata("alloc")
-	obj.load("movie", file, (obj.frame-0.5+1)/obj.framerate)
-	local af, _w, _h = obj.getpixeldata("alloc")
-	ksa.deinterlace_temporal(dest, bf, af, w, h, XOR(obj.frame%2, obj.track1))
-elseif ( obj.track0 == 4 ) then -- 時間方向補間 with 二重化
-	obj.load("movie", file, (obj.frame-0.5-1)/obj.framerate)
-	local past, _w, _h = obj.getpixeldata("alloc")
-	obj.load("movie", file, (obj.frame-0.5+1)/obj.framerate)
-	local future, _w, _h = obj.getpixeldata("alloc")
-	ksa.deinterlace_ghost(dest, past, future, w, h, XOR(obj.frame%2, obj.track1))
-else
-	-- 何もしない
-end
-
-obj.putpixeldata(dest)
-
-@加減速入退場
---track0:入場,0,10,1,0.01
---track1:退場,0,10,1,0.01
---track2:X,-10000,10000,0,1
---track3:Y,-10000,10000,0,1
-
-local t0, t1, dx, dy = obj.track0, obj.track1, obj.track2, obj.track3
-local r = (t0-obj.time)/t0
-if ( 0 < r ) then
-	r = r*r
-	obj.ox = obj.ox + dx*r
-	obj.oy = obj.oy + dy*r
-end
-local r = (t1-obj.totaltime+obj.time)/t1
-if ( 0 < r ) then
-	r = r*r
-	obj.ox = obj.ox + dx*r
-	obj.oy = obj.oy + dy*r
-end
-
-@点滅入退場
---track0:入場,0,100,5,1
---track1:退場,0,100,5,1
---track2:点長,1,1000,2,1
---track3:滅長,1,1000,2,1
-
-local on, off = obj.track2, obj.track3
-local onoff = on + off
-local t0, t1 = onoff*obj.track0, onoff*obj.track1
-
-local r = obj.frame
-if ( r < t0 ) then
-	if ( on <= ( r % onoff ) ) then
-		obj.alpha = 0.0
-	end
-end
-
-local r = obj.totalframe - r - 1
-if ( r < t1 ) then
-	if ( on <= ( r % onoff ) ) then
-		obj.alpha = 0.0
-	end
-end
-
-@スクロール
---track0:再生位置,0,100,0,0.01
---track1:縦/横,0,1,0,1
---track2:表示長,0,10000,100,1
---track3:拡大,0,1,0,1
-
-local ksa = require("ksa")
-local pos = obj.track0/100.0
-local vh = obj.track1==0
-local size = obj.track2
-local extend = obj.track3==1
-
-if ( vh ) then
-	if ( extend ) then
-		obj.effect("領域拡張", "上", size, "下", size, "左", 0, "右", 0, "塗りつぶし", 0)
-	end
-	if ( obj.h < size ) then
-		size = obj.h
-	end
-	local cut = obj.h-size
-	local top, bottom = ksa.round(cut*pos), ksa.round(cut*(1-pos))
-	local r = cut - ( top + bottom )
-	if ( bottom < top ) then
-		top = top + r
-	else
-		bottom = bottom + r
-	end
-	obj.effect("クリッピング", "上", top, "下", bottom, "左", 0, "右", 0, "中心の位置を変更", 1)
-else
-	if ( extend ) then
-		obj.effect("領域拡張", "上", 0, "下", 0, "左", size, "右", size, "塗りつぶし", 0)
-	end
-	if ( obj.w < size ) then
-		size = obj.w
-	end
-	local cut = obj.w-size
-	local left, right = ksa.round(cut*pos), ksa.round(cut*(1-pos))
-	local r = cut - ( left + right )
-	if ( right < left ) then
-		left = left + r
-	else
-		right = right + r
-	end
-	obj.effect("クリッピング", "上", 0, "下", 0, "左", left, "右", right, "中心の位置を変更", 1)
-end
