--- ksa/@ksa.tra
@@ -1,30 +1,0 @@
-@‰ÁŒ¸‘¬ˆÚ“®
---param:5
-local i, x = math.modf(obj.getpoint("index"))
-local a = obj.getpoint(i)
-local b = obj.getpoint(i+1)
-if ( param == 0 ) then
-	return a+(b-a)*(3-2*x)*x*x
-elseif ( param == 1 ) then
-	return a+(b-a)*x*x*x(10+3*x*(2*x-5))
-elseif ( param == 2 ) then
-	local x2 = x*x
-	return a+(b-a)*x2*x2*(35-2*x*(42+5*x*(2*x-7)))
-elseif ( param == 3 ) then
-	local s = math.sin(math.pi*x*0.5)
-	return a+(b-a)*s*s
-elseif ( param == 4 ) then
-	local tau = 2*math.pi
-	return a+(b-a)*(x-math.sin(tau*x)/tau)
-elseif ( param == 5 ) then
-	local px = math.pi*x
-	local s = math.sin(px*0.5)
-	s = s*s
-	return a+(b-a)*(2+math.cos(px))*s*s
-else
-	if ( x < 0.5 ) then
-		return a+(b-a)*2*x*x
-	else
-		return a-(b-a)*(1+2*(x-2)*x)
-	end
-end
