+++ src/extract_functions.sh
@@ -1,8 +1,8 @@
 #!bash
 
-: > functions.c
+: > functions.cpp
 cat ksa_ext.cpp | while read line; do
-	if [[ ${line} =~ ^ksa_(.+)\(lua_State' '\*L\)$ ]]; then
-		echo "	{ \"${BASH_REMATCH[1]}\", KSA::ksa_${BASH_REMATCH[1]} }," >> functions.c
+	if [[ ${line} =~ ^ksa_(.+)\(SCRIPT_MODULE_PARAM' '\*param\)$ ]]; then
+		echo "	{ L\"${BASH_REMATCH[1]}\", KSA::ksa_${BASH_REMATCH[1]} }," >> functions.cpp
 	fi
 done
