+++ src/makezip.sh
@@ -1,7 +1,8 @@
 #!/bin/bash
 
-mkdir KSA
-cp ../ksa/* ./KSA
-VERSION=$(grep '#define VERSION' version.h | sed -E 's/.*"([^"]+)".*/\1/')
-zip -r "../$1.${VERSION}.zip" KSA
-rm -rf KSA
+mkdir Script
+mkdir ./Script/KSA
+cp ../ksa2/* ./Script/KSA
+VERSION=$(grep '#define VERSION' version.hpp | sed -E 's/.*"([^"]+)".*/\1/')
+zip -r "../$1.${VERSION}.au2pkg.zip" Script
+rm -rf Script
