--- src/version.h
@@ -1,1 +1,0 @@
-#define VERSION "1.4.6"
