--- src2/ksa_ext.def
@@ -1,6 +1,0 @@
-LIBRARY ksa_ext
-
-EXPORTS
-	GetScriptModuleTable @1
-	InitializePlugin @2
-	UninitializePlugin @3
